refactor: :truck: moved career cononents into its own folder
